// src/runtime/interfaces/apiIA.interface.ts
export interface RequestIA {
    userInput: string;
}

export interface ResponseIA {
    direccion_original: string;
    direccion_salida: string;
    via_principal: string;
    via: string;
    letra_via: string;
    prefijo_bis: string;
    letra_bis: string;
    via_generadora: string;
    letra_via_generadora: string;
    prefijo_bis_generadora: string;
    letra_bis_generadora: string;
    numero_placa: string;
    cuadrante: string;
    complemento: string;
}